Make tests work by setting the correct collation

+++ images/remove_images.sh
@@ -4,9 +4,27 @@ then
   echo "$0: you need to set docker user via DOCKER_USER=username"
   exit 1
 fi
-docker image rm -f "${DOCKER_USER}/devstats"
-docker image rm -f "${DOCKER_USER}/devstats-minimal"
-docker image rm -f "${DOCKER_USER}/devstats-grafana"
-docker image rm -f "${DOCKER_USER}/devstats-patroni"
-docker image rm -f "${DOCKER_USER}/devstats-test"
-docker system prune -f
+if [ -z "$SKIP_FULL" ]
+then
+  docker image rm -f "${DOCKER_USER}/devstats"
+fi
+if [ -z "$SKIP_MIN" ]
+then
+  docker image rm -f "${DOCKER_USER}/devstats-minimal"
+fi
+if [ -z "$SKIP_GRAFANA" ]
+then
+  docker image rm -f "${DOCKER_USER}/devstats-grafana"
+fi
+if [ -z "$SKIP_PATRONI" ]
+then
+  docker image rm -f "${DOCKER_USER}/devstats-patroni"
+fi
+if [ -z "$SKIP_TEST" ]
+then
+  docker image rm -f "${DOCKER_USER}/devstats-test"
+fi
+if [ -z "$SKIP_PRUNE" ]
+then
+  docker system prune -f
+fi
